Fix for ClusterRoleBinding without a 'subjects' field
ClusterRoleBindings should have the field `subjects`. 
But it is possible to create ClusterRoleBinding wihtout `subjects` field such as the default ClusterRoleBinding `system:node` from kubeadm installlation or by Minikube.
The code didn't verified it the `subjects` field exist and therefore throw an exception.
The fix now verify the this field exist before checking it.