Implement KUBISCAN_{VOLUME,CONFIG_BACKUP}_PATH var
Allow configuration of KubiScan paths through some environment variables
to improve application portability as not ever use-case will allow for
or guarantee the exclusive use of paths such as /tmp.

Furthermore the hardcoded usage of /KubiScan is not quite FHS-friendly
and as such some may opt for the more FHS-compliant /tmp/kubiscan path
or another path that allows for non-sharable write-access.