Add KUBISCAN_CONFIG_PATH to bypass Docker checks
The CONF_PATH variable and the Docker checks provide for a rather
complicated setup. Allowing the user to be explicit about the absolute
path to the configuration file will allow us to skip the Docker checks
and the logic therein and even allow a end-user to to mount the
configuration as a read-only volume. This would be a practical setup for
use in CI/CD pipelines.